refactor(core): sposta normalize_encoding() da csv_read.py a io.py con shim backward compat
